Proyecto finalizado - Resta chequear errores

+++ fulanos/views.py
@@ -1,40 +1,100 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Producto, Usuario, Pedido
-from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
+from .models import Producto, Usuario, Pedido, Avatar
+from django.contrib import messages
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm, UserChangeForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
-from .forms import UsuarioFormulario, ProductoFormulario, BusquedaProductoFormulario
+from .forms import UsuarioFormulario, ProductoFormulario, BusquedaProductoFormulario, ContactForm, UserEditForm, AvatarFormulario
 from django.views.generic import ListView
 from django.views.generic.detail import DetailView
 from django.views.generic.edit import DeleteView, UpdateView, CreateView
+from django.core.mail import BadHeaderError, send_mail
 
 # Usuario
 def crear_usuario(request):
     if request.method == 'POST':
-        formulario = UsuarioFormulario(request.POST)
+        formulario = UserCreationForm(request.POST)
         if formulario.is_valid():
-            formulario.save()
-            return redirect('ListaUsuarios')  
+            user = formulario.save()
+            messages.success(request, f'Usuario {user.username} creado exitosamente.')
+            return redirect('usuario_creado') 
     else:
-        formulario = UsuarioFormulario()
+        formulario = UserCreationForm()
+
     return render(request, 'usuario_formulario.html', {'formulario': formulario})
+
+def usuario_creado(request):
+    return render(request, 'usuario_creado.html')
+
+def es_administrador(user):
+    if not user.is_authenticated:
+        return False
+    return user.is_staff
+
+def no_autorizado(request):
+    return render(request, 'no_autorizado.html', {})
+
+@login_required()
+def agregar_avatar(req):
+
+  if req.method == 'POST':
+
+    mi_formulario= AvatarFormulario(req.POST, req.FILES)
+    if mi_formulario.is_valid():
+
+      data = mi_formulario.cleaned_data
+      avatar = Avatar(user=req.user, imagen=data["imagen"])
+      avatar.save() 
+      return render(req, "inicio.html", { "mensaje": f"Avatar creado correctamente!"})
+
+    else:
+      return render(req, "agregar_avatar.html", { "mi_formulario": mi_formulario })    
+
+  else:
+
+    mi_formulario = AvatarFormulario()
+    return render(req, "agregar_avatar.html", { "mi_formulario": mi_formulario })     
 
 def lista_usuarios(request):
     usuarios = Usuario.objects.all()  
     return render(request, 'lista_usuarios.html', {'usuarios': usuarios})
 
-def editar_usuario(request, id):
-    usuario = Usuario.objects.get(id=id)
-    if request.method == 'POST':
-        formulario = UsuarioFormulario(request.POST, instance=usuario)
-        if formulario.is_valid():
-            formulario.save()
-            return redirect('lista_usuarios')
-    else:
-        formulario = UsuarioFormulario(instance=usuario)
-    return render(request, 'usuario_formulario.html', {'formulario': formulario})
+@login_required
+def editar_perfil(request):
+    avatar_url = None
+    usuario = request.user
 
+    # Obtener el avatar existente si lo hay
+    try:
+        avatar = Avatar.objects.get(user=usuario)
+        avatar_url = avatar.imagen.url if avatar.imagen else None
+    except Avatar.DoesNotExist:
+        avatar = None
+
+    if request.method == 'POST':
+        # Aquí se debe pasar request.POST al formulario
+        mi_formulario = UserEditForm(request.POST, instance=usuario)  # Asegúrate de pasar request.POST
+
+        # Manejo de la carga del avatar
+        if request.FILES.get('avatar'):
+            if avatar:
+                avatar.imagen = request.FILES['avatar']  # Actualizar imagen existente
+            else:
+                avatar = Avatar(user=usuario, imagen=request.FILES['avatar'])  # Crear nuevo avatar
+            avatar.save()  # Guardar el avatar
+
+        if mi_formulario.is_valid():  # Asegúrate de validar el formulario
+            mi_formulario.save()  # Guardar los cambios del formulario
+            messages.success(request, 'Los datos han sido actualizados con éxito.')  # Añadir mensaje de éxito
+            return redirect('inicio')  # Redirigir a la página de inicio
+    else:
+        mi_formulario = UserEditForm(instance=usuario)  # Formulario para GET
+
+    return render(request, 'editar_perfil.html', {
+        'mi_formulario': mi_formulario,
+        'avatar_url': avatar_url,
+    })
 class UsuarioList(ListView):
     model = Usuario
     template_name = 'lista_usuarios.html'
@@ -45,8 +105,6 @@ class UsuarioCreate(CreateView):
     template_name = 'usuario_formulario.html'
     fields = ['nombre', 'apellido', 'email']
     success_url = '/fulanos'
-
-from django.shortcuts import redirect
 
 def login_view(req):
     if req.method == 'POST':
@@ -98,15 +156,19 @@ def logout_view(request):
     return render(request, "logout.html")
 
 # Producto
+@login_required
 def crear_producto(request):
+    if not request.user.is_staff:
+        return redirect('no_autorizado')  # Redirige a la página personalizada
+
     if request.method == 'POST':
-        formulario = ProductoFormulario(request.POST, request.FILES)  
+        formulario = ProductoFormulario(request.POST, request.FILES)
         if formulario.is_valid():
             formulario.save()
             return redirect('ListaProductos')
     else:
         formulario = ProductoFormulario()
-    
+
     return render(request, 'producto_formulario.html', {'formulario': formulario})
 
 @login_required
@@ -114,23 +176,37 @@ def lista_productos(request):
     productos = Producto.objects.all()
     return render(request, 'lista_productos.html', {'productos': productos})
 
-def editar_producto(request, id):
-    producto = Producto.objects.get(id=id)
+@login_required
+def editar_producto(request, id=None):
+    if not request.user.is_staff:
+        return redirect('no_autorizado')  # Redirige a la página personalizada
+
+    producto = get_object_or_404(Producto, id=id) if id else None
+
     if request.method == 'POST':
-        formulario = ProductoFormulario(request.POST, instance=producto)
+        formulario = ProductoFormulario(request.POST, request.FILES, instance=producto)
         if formulario.is_valid():
             formulario.save()
             return redirect('ListaProductos')
     else:
         formulario = ProductoFormulario(instance=producto)
-    return render(request, 'producto_formulario.html', {'formulario': formulario})
 
+    return render(request, 'producto_formulario.html', {
+        'formulario': formulario,
+        'producto': producto,  
+    })
+
+@login_required
 def eliminarproducto(request, id):
+    if not request.user.is_staff:
+        return redirect('no_autorizado')  # Redirige a la página personalizada
+
     producto = get_object_or_404(Producto, id=id)
     if request.method == 'POST':
         producto.delete()
         return redirect('ListaProductos')  
     return render(request, 'confirmar_eliminar.html', {'producto': producto})
+
 
 def detalle_producto(request, id):
     producto = get_object_or_404(Producto, id=id)
@@ -143,7 +219,7 @@ class ProductoDetail(DetailView):
   context_object_name = 'producto'
 
 # Pedido
-class PedidoList(ListView):
+class PedidoList(LoginRequiredMixin, ListView):
     model = Pedido
     template_name = 'pedido_list.html'
     context_object_name = 'pedidos'
@@ -171,8 +247,13 @@ class PedidoDelete(DeleteView):
     success_url = '/fulanos'
         
 # Inicio
-def inicio(request):
-    return render(request, 'inicio.html')
+def inicio(req):
+    try:
+        avatar = Avatar.objects.filter(user=req.user.id).first()  
+        return render(req, "inicio.html", {'url': avatar.imagen.url})
+
+    except:
+        return render(req, "inicio.html", {})
 
 # Buscar producto
 def busqueda_producto(request):
@@ -190,6 +271,20 @@ def buscar_producto(request):
 #About
 def acerca_de_mi(request):
     return render(request, 'acerca_de_mi.html')
+
+#Contacto
+def contacto(request):
+    if request.method == 'POST':
+        form = ContactForm(request.POST)
+        if form.is_valid():
+            # Aquí no se envía el correo, solo se muestra el mensaje de éxito, a desarrollar proximamente.
+            messages.success(request, "¡Tu mensaje ha sido enviado exitosamente!")
+            return redirect('contacto') 
+
+    else:
+        form = ContactForm()
+
+    return render(request, 'contacto.html', {'form': form})
 
 #404 error
 def error_404_view(request, exception):
